refactor: extract options per view
